Removed Unnecessary NetworkX Dependency
I decided to move away from the use of NewtorkX as a base class for the graph generation as I was only using its most basic tenets. By doing this I am able to more accurately create and track data generated within the shapes (such as position, size, edge weight, and colour) as well as ease the use of my code as an imported library.